Create new component for assessment actions row
Get tag create working again
